Adding javascript powered balance input button
This provides a shortcut to easily enter in the balance to be paid
rather than manually enter the amount to be paid and selecting the
tender type. With this, you can simply hit the balance button and then
hit the tender type to enter the payment for a transaction. Currently
this is mapped to Ctrl+9 in js in the header of input.php